+++ src/Bootstrap4Jumbotron.php
@@ -120,7 +120,7 @@ class Bootstrap4Jumbotron extends AbstractHelper
     }
 
     /**
-     * Worker method for the view helper, generates the HTML, the method id private so that we
+     * Worker method for the view helper, generates the HTML, the method is private so that we
      * can echo/print the view helper directly
      *
      * @return string